feat: merge Profile into Settings, remove fake landing/settings content
- Remove the fabricated "trusted by" companies (CapitalView, GrowthFlow,
  StackWise) from the landing page
- Drop the DBS · SGFinDex connection row and the non-functional
  "Require Face ID for vault" toggle from Settings
- Merge the Profile page into Settings as one sectioned page (General,
  Security, Appearance, Privacy & vault, Data & connections); /profile
  now redirects to /settings and the sidebar Profile link is gone
- Update Settings tests to cover the merged sections and assert the
  removed items no longer render
