+++ frontend/src/components/sidebar.tsx
@@ -4,7 +4,6 @@ import {
     ArrowRightLeft,
     PieChart,
     History,
-    UserCircle,
     Wallet,
     Landmark,
     Users,
@@ -16,7 +15,6 @@ import {
 } from "lucide-react";
 import SidebarButton from "./sidebarButton";
 import { useAuth } from "../lib/AuthContext";
-import { useNavigate } from "react-router";
 
 function Sidebar() {
     // Pull the authentication state from your context
@@ -48,7 +46,6 @@ function Sidebar() {
                     {/* Conditionally render based on auth state */}
                     {isAuthenticated ? (
                         <>
-                            <SidebarButton text="Profile" href="/profile" icon={<UserCircle />} />
                             <Form method="post" action="/logout">
                                 <button type="submit" className="flex w-full items-center gap-3 rounded-lg px-3 py-2 text-sm font-medium transition-colors text-base-600 hover:bg-base-100 hover:text-base-900 dark:text-base-400 dark:hover:bg-base-900 dark:hover:text-base-100">
                                     <LogOut className="h-4 w-4" />
